Started coding full file compression tests

+++ src/format.h
@@ -16,10 +16,10 @@ typedef struct sym_hdr sym_hdr;
 //The memory image of the following struct will be the first thing in a compressed file
 struct elfp_hdr
 {
-	uint64_t sym_offset; //Offset to the symbol table
+	uint32_t sym_offset; //Offset to the symbol table
 	uint32_t num_syms; //Number of entries in symbol table
 
-	uint64_t data_offset; //Offset to compressed data
+	uint32_t data_offset; //Offset to compressed data
 	int64_t num_bits; //Actual size of compressed data in bits
 	size_t comp_size; //Size of the compressed data in file
 	size_t uncomp_size; //Size of decompressed data
